Del opp i søknadsfriststeg og grunnlaguttaksteg
