+++ web/src/test/java/no/nav/foreldrepenger/web/app/tjenester/behandling/søknad/aksjonspunkt/VurderSøknadsfristOppdatererTest.java
@@ -3,10 +3,10 @@ package no.nav.foreldrepenger.web.app.tjenester.behandling.søknad.aksjonspunkt;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
-import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
 
+import javax.enterprise.inject.Any;
+import javax.enterprise.inject.Instance;
 import javax.inject.Inject;
 import javax.persistence.EntityManager;
 
@@ -15,11 +15,9 @@ import org.junit.jupiter.api.Test;
 import no.nav.foreldrepenger.behandling.Søknadsfrister;
 import no.nav.foreldrepenger.behandling.aksjonspunkt.AksjonspunktOppdaterParameter;
 import no.nav.foreldrepenger.behandling.aksjonspunkt.BekreftetAksjonspunktDto;
-import no.nav.foreldrepenger.behandlingskontroll.FagsakYtelseTypeRef;
 import no.nav.foreldrepenger.behandlingslager.behandling.Behandling;
 import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingStegType;
 import no.nav.foreldrepenger.behandlingslager.behandling.Behandlingsresultat;
-import no.nav.foreldrepenger.behandlingslager.behandling.BehandlingsresultatRepository;
 import no.nav.foreldrepenger.behandlingslager.behandling.aksjonspunkt.AksjonspunktDefinisjon;
 import no.nav.foreldrepenger.behandlingslager.behandling.repository.BehandlingRepositoryProvider;
 import no.nav.foreldrepenger.behandlingslager.behandling.søknad.SøknadEntitet;
@@ -28,33 +26,23 @@ import no.nav.foreldrepenger.behandlingslager.behandling.ytelsefordeling.Avklart
 import no.nav.foreldrepenger.behandlingslager.behandling.ytelsefordeling.periode.FordelingPeriodeKilde;
 import no.nav.foreldrepenger.behandlingslager.behandling.ytelsefordeling.periode.OppgittFordelingEntitet;
 import no.nav.foreldrepenger.behandlingslager.behandling.ytelsefordeling.periode.OppgittPeriodeBuilder;
-import no.nav.foreldrepenger.behandlingslager.behandling.ytelsefordeling.periode.OppgittPeriodeEntitet;
 import no.nav.foreldrepenger.behandlingslager.behandling.ytelsefordeling.periode.UttakPeriodeType;
-import no.nav.foreldrepenger.behandlingslager.fagsak.FagsakYtelseType;
 import no.nav.foreldrepenger.behandlingslager.testutilities.behandling.ScenarioMorSøkerForeldrepenger;
 import no.nav.foreldrepenger.behandlingslager.uttak.Uttaksperiodegrense;
 import no.nav.foreldrepenger.behandlingslager.uttak.UttaksperiodegrenseRepository;
 import no.nav.foreldrepenger.dbstoette.CdiDbAwareTest;
-import no.nav.foreldrepenger.domene.ytelsefordeling.YtelseFordelingTjeneste;
 
 @CdiDbAwareTest
-public class VurderSøknadsfristOppdatererTjenesteFPTest {
+public class VurderSøknadsfristOppdatererTest {
 
     @Inject
-    @FagsakYtelseTypeRef(FagsakYtelseType.FORELDREPENGER)
-    private VurderSøknadsfristOppdatererTjenesteFP tjeneste;
+    private @Any Instance<VurderSøknadsfristOppdaterer> tjeneste;
 
     @Inject
     private UttaksperiodegrenseRepository uttaksperiodegrenseRepository;
 
     @Inject
-    private BehandlingsresultatRepository behandlingsresultatRepository;
-
-    @Inject
     private SøknadRepository søknadRepository;
-
-    @Inject
-    private YtelseFordelingTjeneste ytelseFordelingTjeneste;
 
     @Test
     public void skal_oppdatere_behandlingsresultet_med_uttaksperiodegrense(EntityManager em) {
@@ -66,7 +54,7 @@ public class VurderSøknadsfristOppdatererTjenesteFPTest {
         var behandling = byggBehandlingMedYf(em);
 
         // Act
-        tjeneste.oppdater(dto, aksjonspunktParam(behandling, dto));
+        tjeneste.get().oppdater(dto, aksjonspunktParam(behandling, dto));
 
         // Assert
         var uttaksperiodegrense = uttaksperiodegrenseRepository.hent(behandling.getId());
@@ -95,31 +83,13 @@ public class VurderSøknadsfristOppdatererTjenesteFPTest {
         dto.setAnsesMottattDato(nyMottattDato);
 
         // Act
-        tjeneste.oppdater(dto, aksjonspunktParam(behandling, dto));
+        tjeneste.get().oppdater(dto, aksjonspunktParam(behandling, dto));
 
         // Assert
         var uttaksperiodegrense = uttaksperiodegrenseRepository.hent(behandling.getId());
         assertThat(uttaksperiodegrense.getErAktivt()).isTrue();
         assertThat(uttaksperiodegrense.getMottattDato()).isEqualTo(nyMottattDato);
         assertThat(Søknadsfrister.tidligsteDatoDagytelse(uttaksperiodegrense.getMottattDato())).isEqualTo(førsteLovligeUttaksdag);
-    }
-
-    @Test
-    public void skal_oppdatere_mottatt_dato_i_oppgitte_perioder(EntityManager em) {
-        var nyMottattDato = LocalDate.of(2018, 1, 15);
-        var dto = new VurderSøknadsfristDto("begrunnelse", true);
-        dto.setAnsesMottattDato(nyMottattDato);
-
-        var behandling = byggBehandlingMedYf(em);
-        tjeneste.oppdater(dto, aksjonspunktParam(behandling, dto));
-
-        var ytelseFordelingAggregat = ytelseFordelingTjeneste.hentAggregat(behandling.getId());
-        var justertFordelingSortert = ytelseFordelingAggregat.getJustertFordeling().orElseThrow().getPerioder().stream()
-                .sorted(Comparator.comparing(OppgittPeriodeEntitet::getFom))
-                .collect(Collectors.toList());
-        // Skal ikke oppdatere vedtaksperioder
-        assertThat(justertFordelingSortert.get(0).getMottattDato()).isNotEqualTo(nyMottattDato);
-        assertThat(justertFordelingSortert.get(1).getMottattDato()).isEqualTo(nyMottattDato);
     }
 
     @Test
@@ -135,7 +105,7 @@ public class VurderSøknadsfristOppdatererTjenesteFPTest {
         var dto = new VurderSøknadsfristDto("bg", false);
         dto.setAnsesMottattDato(mottattDatoSøknad.plusYears(1));
 
-        tjeneste.oppdater(dto, aksjonspunktParam(behandling, dto));
+        tjeneste.get().oppdater(dto, aksjonspunktParam(behandling, dto));
 
         var uttaksperiodegrense = uttaksperiodegrenseRepository.hent(behandling.getId());
